Add files via upload
Create a priority queue that is based on a heap. Next, complete the implementation of Heap.h and PQType.h.

The driver will prompt the user for input and output files, then run all of the commands from the input file. As this loop processes the input file for commands, it is also updating the output file.